requirements.txt
attrs==25.3.0
audioop-lts==0.2.1
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
filelock==3.18.0
Flask==3.1.1
flatbuffers==25.2.10
fsspec==2025.7.0
gunicorn==23.0.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
lazy_loader==0.4
llvmlite==0.44.0
lxml==6.0.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
onnxruntime==1.22.1
openai-whisper @ git+https://github.com/openai/whisper.git@c0d2f624c09dc18e709e37c2ad90c039a4eb72a2
opencv-contrib-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pdf2image==1.17.0
pillow==11.3.0
platformdirs==4.3.8
pooch==1.8.2
protobuf==6.31.1
pydub==0.25.1
PyMatting==1.1.14
PyMuPDF==1.26.3
PyPDF2==3.0.1
pyreadline3==3.5.4
python-docx==1.2.0
pytube==15.0.0
referencing==0.36.2
regex==2025.7.29
rembg==2.0.67
requests==2.32.4
rpds-py==0.26.0
scikit-image==0.25.2
scipy==1.16.0
setuptools==80.9.0
six==1.17.0
SpeechRecognition==3.14.3
standard-aifc==3.13.0
standard-chunk==3.13.0
sympy==1.14.0
tifffile==2025.6.11
tiktoken==0.9.0
torch==2.7.1
tqdm==4.67.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
yt-dlp==2025.7.21


anyio==4.10.0
blinker==1.9.0
certifi==2025.8.3
chardet==3.0.4
click==8.2.1
colorama==0.4.6
defusedxml==0.7.1
fire==0.7.0
fonttools==4.59.0
fpdf2==2.7.4
googletrans-py==4.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
hstspreload==2025.1.1
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.0
MarkupSafe==3.0.2
numpy==2.2.6
opencv-python-headless==4.12.0.88
packaging==25.0
pdf2docx==0.5.8
pillow==11.3.0
PyMuPDF==1.26.3
pytesseract==0.3.10
python-docx==0.8.11
python-magic==0.4.27
rfc3986==1.5.0
sniffio==1.3.1
termcolor==3.1.0
typing_extensions==4.14.1
Werkzeug==3.1.3
